Implement legacy environment handling
This has slightly changed over the old method of links.  Instead, specify the container name
directly in the environment variable and it will be looked up by inspection and
used if it exists.  This has the advantage of the proxy container *not* failing to start
if the referenced container is missing